requirements.txt
asgiref==3.8.1
autocommand==2.2.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Django==5.1.1
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.9
inflect==7.4.0
inflection==0.5.1
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.itertools==6.4.1
jaraco.text==4.0.0
jwcrypto==1.5.6
more-itertools==10.5.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
pystache==0.6.5
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
setuptools==74.1.2
sqlparse==0.5.1
static==1.1.1
typeguard==4.3.0
typing_extensions==4.12.2
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==2.2.3
user-agents==2.2.0
